Made Multiple changes to the code, implemented the smart functionality.
up to ibrahim to beautify now.

Encountered a major issue ~ that the user needs to have a phpmysql server on their pc to run, in order for the application to work.

I asked an AI agent what options we have  and it basically spat out the readME file, so please take a look

we obviously need to shortern it and get to work in choosing which one is the best